% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2010.R
\docType{data}
\name{b2010}
\alias{b2010}
\title{Base de dados de 2010}
\format{
Uma base de dados com 18 linhas e 3 colunas
\describe{
\item{faixa.etaria}{faixas etarias}
\item{Pessoas.mortas}{Pessoas que morreram em 2010, dividido por faixa etaria}
\item{Pessoas.vivas}{Pessoas vivas em 2010, dividido por faixa etaria}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
b2010
}
\description{
Uma base de de dados que descreve as pessoas vivas e as pessoas que morreram por faixa etaria
}
\keyword{datasets}
