% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\docType{data}
\name{covid}
\alias{covid}
\title{Covid 19 no Brasil}
\format{
Uma base de dados com 2727 linhas e 9 colunas
\describe{
\item{X}{contador de anotaçoes}
\item{region}{regiao}
\item{state}{estado ao qual o dado diz respeito}
\item{newCases}{numero de novos casos registrados}
\item{accumCases}{numero de casos registrados ate determinado momento}
\item{newDeaths}{numero de novas mortes registrados}
\item{accumDeaths}{numero de mortes registradas ate determinado momento}
\item{date}{data a qual os dados dizem respeito}
\item{mortality}{mortalidade}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
covid
}
\description{
Uma base de de dados que contem dados de todo o Brasil sobre a covid 19
}
\keyword{datasets}
