% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelas.R
\name{life.table}
\alias{life.table}
\title{tabela de mortalidade}
\usage{
life.table(x, nDx, nKx)
}
\arguments{
\item{x}{faixas etarias}

\item{nDx}{pessoas que morreram nos periodos estudados por faixa etaria}

\item{nKx}{pessoas vivas nos periodos estudados por faixa etaria}
}
\value{
retorna uma tabela de mortalidade
}
\description{
criamos uma tabela de mortalidade a partir da variaveis x,nDx e nKx.
}
\seealso{
\url{https://r-pkgs.org/whole-game.html#whole-game-document}
}
\author{
Victor Prates
}
