% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipotese_normal.R
\name{normalidade}
\alias{normalidade}
\alias{normal}
\title{teste hipotese utilizando a distribuição normal}
\usage{
normal(a, mp, var, conf.level = 0.95, tipo)
}
\arguments{
\item{a}{amostra}

\item{mp}{media populacional}

\item{var}{variancial populacional}

\item{conf.level}{intervalo de confiança}

\item{tipo}{qual o tipo de teste que será realizado :
"diferente" testa se a media amostral é diferente da media populacional (realiza um teste bilateral)
"menor" testa se a  media amostral é menor que a media populacional (realiza um teste unilateral esquerdo)
"maior" testa se a  media amostral é maior que a media populacional (realiza um teste unilateral direito)}
}
\value{
retorna a medida a ser tomada de acordo com a analise
}
\description{
utilizamos calculos para definir se uma hipotese nula deve ou não ser rejeitada,sabendo a media e variancia populacional.
}
\seealso{
\url{https://r-pkgs.org/whole-game.html#whole-game-document}
}
\author{
Victor Prates
}
